+++ Toris/Assets/Scripts/Items/Interfaces/IEquipable.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace OutlandHaven.Inventory
+{
+    public interface IEquipable
+    {
+        EquipmentSlot TargetSlot { get; }
+        float StrengthBonus { get; }
+        float DefenceBonus { get; }
+    }
+}