+++ Toris/Assets/Scripts/Items/Interfaces/IUsable.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace OutlandHaven.Inventory
+{
+    public interface IUsable
+    {
+        bool TryUse(PlayerConsumableController consumableController, InventoryManager inventoryManager, InventorySlot slot);
+    }
+}