+++ Toris/Assets/Scripts/UIToolkit/UI/Controllers/HudScreenController.cs
@@ -10,6 +10,7 @@ namespace OutlandHaven.UIToolkit
         [SerializeField] private VisualTreeAsset _hudMainTemplate; // <--- Drag HUD.uxml here
         [SerializeField] private VisualTreeAsset _buttonTemplate;
         [SerializeField] private VisualTreeAsset _slotTemplate;
+        [SerializeField] private UIInventoryEventsSO _uiInventoryEvents;
         [SerializeField] private GameSessionSO _gameSession;
         [SerializeField] private UIEventsSO _uiEvents;
         [SerializeField] private InventoryManager _potionInventory;
@@ -43,7 +44,7 @@ namespace OutlandHaven.UIToolkit
             }
 
             // 2. Pass the INSTANCE to the View
-            _view = new HUDView(hudInstance, _playerHudBridge, _uiEvents, _buttonTemplate, _slotTemplate);
+            _view = new HUDView(hudInstance, _playerHudBridge, _uiEvents, _uiInventoryEvents, _buttonTemplate, _slotTemplate);
             _view.Initialize();
             _view.Setup(_potionInventory);
 
@@ -51,6 +52,7 @@ namespace OutlandHaven.UIToolkit
             _uiManager.RegisterView(_view, ScreenZone.HUD);
         }
 
+#if UNITY_EDITOR
         private void OnValidate()
         {
             if (_hudMainTemplate == null)
@@ -59,10 +61,13 @@ namespace OutlandHaven.UIToolkit
                 Debug.LogWarning($"[UI/Inventory] <color=red>{name}</color> missing Button Template", this);
             if (_slotTemplate == null)
                 Debug.LogWarning($"[UI/Inventory] <color=red>{name}</color> missing Slot Template", this);
+            if (_uiInventoryEvents == null)
+                Debug.LogError($"<b><color=red>[HUD]</color></b> missing <b>UIInventoryEventsSO</b> on GameObject: <b>{name}</b>", this);
             if (_uiEvents == null)
                 Debug.LogWarning($"[UI/Inventory] <color=red>{name}</color> missing UI Events SO", this);
             if (_gameSession == null)
                 Debug.LogWarning($"[UI/Inventory] <color=red>{name}</color> missing Game Session SO", this);
         }
+#endif
     }
 }